fix(terminal): match PAM MOTD sources
Ultraworked with [omo](https://github.com/code-yeongyu/oh-my-openagent)
