+++ .tegami/feat-session-motd.md
@@ -7,6 +7,7 @@ packages:
 ## Show the login message of the day
 
 Terminal sessions now display the message of the day before the login shell
-starts, matching an interactive `ssh` login. The banner follows `pam_motd`'s
-default file and directory precedence, honors `.hushlogin`, and never reaches
-the shell as input. Reconnecting to the same session does not print it again.
+starts, matching an interactive `ssh` login. The banner includes Ubuntu's
+generated dynamic message followed by `pam_motd`'s default file and directory
+sources, honors `.hushlogin`, and never reaches the shell as input.
+Reconnecting to the same session does not print it again.
